Tests clean-up Modified Service/DataProvider not found error to be more Test Friendly
