+++ Tests/RealityToolKit.ServiceFramework.Tests.asmdef
@@ -7,7 +7,8 @@
         "GUID:13703f41b24bb904cb2305abe6317e3d",
         "GUID:9753fcbb5b1feaf459f435ac95e51baa",
         "GUID:b2d046948d6452a4b8485efc9ce0f88c",
-        "GUID:2a3f0ca4e21332c44bfdce311ea8943e"
+        "GUID:2a3f0ca4e21332c44bfdce311ea8943e",
+        "GUID:575f10fdf9929b24db3492e5c8f92838"
     ],
     "includePlatforms": [
         "Editor"
