+++ Tests/TestData/Profiles/TestService1Profile.cs
@@ -3,6 +3,6 @@ using RealityToolkit.ServiceFramework.Interfaces;
 
 namespace RealityToolkit.ServiceFramework.Tests.Profiles
 {
-    internal class TestService1Profile : BaseServiceProfile<IServiceDataProvider>
+    public class TestService1Profile : BaseServiceProfile<IServiceDataProvider>
     { }
 }
