+++ Tests/TestData/Providers/TestDataProvider1.cs
@@ -4,16 +4,23 @@ using RealityToolkit.ServiceFramework.Definitions;
 using RealityToolkit.ServiceFramework.Interfaces;
 using RealityToolkit.ServiceFramework.Providers;
 using RealityToolkit.ServiceFramework.Tests.Interfaces;
+using UnityEngine;
 
 namespace RealityToolkit.ServiceFramework.Tests.Providers
 {
     [System.Runtime.InteropServices.Guid("407D379E-3351-4B2D-9C88-1B54C42B5554")]
-    internal class TestDataProvider1 : BaseServiceDataProvider, ITestDataProvider1
+    public class TestDataProvider1 : BaseServiceDataProvider, ITestDataProvider1
     {
         public const string TestName = "Test Data Provider 1";
 
         public TestDataProvider1(string name = TestName, uint priority = 1, BaseProfile profile = null, IService parentService = null)
             : base(name, priority, profile, parentService)
         { }
+
+        public override void Initialize()
+        {
+            base.Initialize();
+            Debug.Log($"{TestName} is Initialised");
+        }
     }
 }
