+++ Tests/TestData/RealityToolkit.ServiceFramework.TestData.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 575f10fdf9929b24db3492e5c8f92838
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
