+++ Tests/TestData/Services/TestService1.cs
@@ -3,16 +3,23 @@
 using RealityToolkit.ServiceFramework.Definitions;
 using RealityToolkit.ServiceFramework.Services;
 using RealityToolkit.ServiceFramework.Tests.Interfaces;
+using UnityEngine;
 
 namespace RealityToolkit.ServiceFramework.Tests.Services
 {
     [System.Runtime.InteropServices.Guid("80B2B43B-F18B-4E68-A9AB-505290D31110")]
-    internal class TestService1 : BaseServiceWithConstructor, ITestService
+    public class TestService1 : BaseServiceWithConstructor, ITestService1
     {
         public const string TestName = "Test Service 1";
 
         public TestService1(string name = TestName, uint priority = 0, BaseProfile profile = null)
             : base(name, priority)
         { }
+
+        public override void Initialize()
+        {
+            base.Initialize();
+            Debug.Log($"{TestName} is Initialised");
+        }
     }
 }
