+++ Tests/Tests/DataProvider Registration Tests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3743655d1cff48e42af84424ad4a03de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
