+++ Tests/Tests/DataProvider Retrieval Tests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ed1be826ba03e294bab6f1a9fa5b108a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
