+++ Tests/Tests/DataProvider UnRegistration Tests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db21f37161514e748b78ee0b950ed868
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
