+++ Tests/Tests/Service Enable Tests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ee061873e66e4942abc9efb7f6e654a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
