+++ Tests/Tests/Service Locator Tests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5dcbf462a93bbf24ca835f4ba5efc382
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
