+++ Tests/Tests/Service Registration Tests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 63598453c3b25bc4597cc7933af5bb31
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
