+++ Tests/Tests/Service Retrieval Tests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07442ca396662bc438d6b00d87eca4c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
