+++ Tests/Tests/Service UnRegistration Tests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d688237ff95cfeb41bce9419b64935fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
